combine info fields and image uploads into one component
